+++ packages/koop-core/test/fixtures/fake-auth/index.js
@@ -1,0 +1,15 @@
+
+function fakeAuth () {
+
+  return {
+    type: 'auth',
+    getAuthenticationSpecification: function () {
+      return function () { }
+    },
+    authenticate: function () {},
+    authorize: function () {}
+  }
+}
+
+
+module.exports = fakeAuth